MessagePack: dissect more types+other improvements
Add support for unhandled msgpack data formats:

* `map 16` and `map 32` (expand from `fixmap` support)
* `ext 8`/`ext 16`/`ext 32` (expand from `fixext` support)
* `bin 8`/`bin 16`/`bin 32` (replicate from string type support)

Add a `range_string` lookup for describing the format byte. Create a
tree entry for the format byte in more places. Remove the `value_string`
that was used for this purpose but only on extension types.

Correctly handle negative `fixint`s by representing them as negative
numbers.

Fix an inconsistency in string-type handling that falsely stated
`fixstr`s had a length of 1 byte.

Fix a logic error in map and array handling which treated the number of
elements in the structure as being the number of bytes used to represent
the number of elements.

Replace `tvb_get_ptr()`+`bytes_to_hexstr()` with `tvb_bytes_to_str()` in
extension-type handling.

Change field type for `float 32` from `FT_STRING` to `FT_FLOAT`. Create
separate field for `float 64` as `FT_DOUBLE`.

Use `ENC_UTF_8` instead of `ENC_ASCII` for messagepack strings in
accordance to the spec.

Change some of the generated labels for consistency, and use them more
consistently.

Possible future enhancements:

* More detailed breakdown of types and lengths (need to consult e.g.
  `packet-protobuf.c` for precedent)
* Dissector table for positive extension types
* Handle extension type -1 (timestamps)
* Expert info on other negative (and thus reserved) extension types
* More graceful handling of the "never used"/deprecated `0xc1` type
* More hardening in general

For testing I generated
<https://gitlab.com/wireshark/wireshark/-/wikis/uploads/ecaf7e0e2555d472607ea4c7323e24a5/msgpack-generated.pcap>