ftypes: Use C23 compatible checked arithmetic
Add a header-only library for C23 checked integer arithmetic from
https://github.com/jart/jtckdint

Use it in the ftypes handling instead of the implementations in
safe-math. Note that the return value is opposite (true means
overflow occurred).

C23 stdckdint.h (and jtckdint.h) do not include implementations
for div, mod, and unary negation, where division by zero and,
for signed integers, cases like INT_MIN / -1 which overflow from
INT_MAX + 1 for 2's complement. Those cases have to be handled
separately. Note that the divide by zero case, unlike the others,
can't really have a "this is what the answer is if you just want
the result to wrap around," though that doesn't matter for the
ftypes/dfilter.

Ping #20144