tap-iostat: add reset and finish routines
Move code which freed the memory state of the tap out of `iostat_draw()`
and into a new `iostat_finish()` routine. Add code to this routine which
frees the doubly-linked list of stat items, which was previously leaked.

Add `iostat_reset()` function to reset all statistics when a new file is
loaded. Both this and `iostat_finish()` are only registered as callbacks
when registering the first iostat tap, to match that `iostat_draw()` was
also only registered once in this same way.

Move `last_relative_time` from a static global to the per-instance tap
data.

Ping #20432