nstime: Don't overflow in nstime_cmp
nstime_cmp can overflow unnecessarily and give the opposite result
in cases where a time_t is bigger than an int (e.g., 64-bit and 32-bit,
as on many platforms) and two times differ by more than INT_MAX.

Just return -1, 0, and 1. If it's truly useful to return something like
"the difference in the time in seconds, rounded away from zero," we
could do that with C23 checked integer arithmetic. (However, the old
implemention couldn't be counted on to do that, and it would return
the difference in the nanoseconds if the seconds were equal.)

Fix #16463