protobuf: Use col_append_sep_fstr
col_append_sep_fstr adds the separator if and only if there's existing
text in the column, so just use that instead of reinventing it.

It shouldn't be necessary to clear COL_INFO only if COL_PROTOCOL is
empty, because an empty Protocol column should imply an empty Info
column.