epan: various time-related cleanups and impreovements.
Have routines to calculate, for a frame:

* the time relative to the first frame in the capture;

* the time relative to the previous reference frame or, if there is
  none, the first frame in the capture;

* the time relative to the previous captured frame;

* the time relative to the previous displayed frame.

That centralizes the determination of which frame that is, so that if we
want to change the rule for determining that (e.g., if we decide to
change to "relative to the previous XXX frame *that has a time stamp*),
we only need to change it in one place.

When computing relative times, make sure both the current frame and the
frame from which a relative time is being computed exist, and, if not,
indicate that the relative time is missing (even if we set its value to
zero).

Don't display relative times (whether relative to the first frame, the
most recent reference frame, the previous captured frame, or the
previous displayed frame) that don't, in fact, exists; omit the protocol
tree field, and show a blank in the time column.

Remove a ws_debug message in epan_get_frame_ts() if we can't get the
frame time stamp, as that may just mean it didn't *have* a time stamp.

When sorting columns by time, make non-existent times - which display as
blank - sort to the top.

Note, in comments, some issues found while doing this.

Update tests as appropriate for the first packet not having a "time
since previous packet" value.  This includes adding some packets to
nfs.pcap, from the same file whence they originally came (with times
shifted), with additional time shifting to make all tests doable.