msgpack: assert on invalid type bytes
Add `DISSECTOR_ASSERT_NOT_REACHED()` to functions handling the
various messagepack types, to reflect that they will only be called with
specific values of the `type` byte. This should have no functional or
performance effect, but it will save a developer having to scroll down
to `dissect_msgpack_object()` to see how the type function is called.