wiretap: add a register routine to the pcapng NFLX custom code.
Currently empty, but the goal is to have a mechanism for custom block
and option code. and local block and option code, to register to be
called by the pcapng code, so that, in most if not all case, *no*
changes to other Wireshark code is needed to add support for those.