cpu_info: improve the CPU name when running under Rosetta 2.
Add a routine to determine whether you're running x86-64 code under
Rosetta 2 on an ARM64 machine and, if you are, note after the CPU name -
which will refer to the Apple silicon CPU name - that the program is
running under Rosetta (so that the mention of SSE 4.2, which Rosetta 2
supports, doesn't look as if Arm have their own "SSE" extension).

See #17294 comment
https://gitlab.com/wireshark/wireshark/-/issues/17294#note_2491249686
*et seq*.

This could, in principle, be extended to handle Microsoft's x86-64 ->
ARM64 translator and any other such translators, if there's a way to
detect from within the Matrix that you're in the Matrix.


(cherry picked from commit 8007321b3cfc73a9557192fd3e257b5deb1eb4e3)

Co-authored-by: Guy Harris <gharris@sonic.net>