pcapng: use the three byte-order values when writing options.
Have pcapng_write_options() take a pcapng_opt_byte_order_e argument, and
have the option-specific write function take that argument as well.
For string options and custom options, change the byte order if the
byte-order setting is "use little-endian" or "use big-endian" rather
than "use the order of the host writing the file". (Other options aren't
used in custom blocks, which are the only blocks where "use
little-endian" or "use big-endian" would be used for options.)

Also, pass an err_info argument, as an invalid byte-order value will
cause the writer to return WTAP_ERR_INTERNAL and an error message that
indicates what the problem is.

This means we can get rid of the "use_little_endian" variable and
structure member, as that's just a roundabout way of treating options in
custom blocks differently fro options in standard blocks. We now do that
by having the routine that writes a block pass the appropriate
byte-order value to pcapng_write_options().

This also causes the option type and length for the comment option to be
written in the same byte order as custom options, as indicated by
section 5.4 "Endianness Issues" of the pcapng Internet-Draft

https://datatracker.ietf.org/doc/html/draft-ietf-opsawg-pcapng#section-5.4

> Therefore, the vendor should either encode the Custom Data of their
> Custom Blocks and Custom Options, the Option Type and Option Length
> fields of options in Custom Blocks, and the PEN field of Custom
> Options in Custom Blocks in a consistent manner, such as always in
> big-endian or always in little-endian format, regardless of the host
> platform's endianness, or should encode some flag in the Custom Data
> payload to indicate in which endianness the rest of the payload is
> written.

Fix other things that showed up (format %d for an unsigned value,
copy-and-paste error in a comment).