HTTP/3: Clear the Info column as intended
Inside dissect_http3, the http3 dissector has already been called and
so "http3" is always inside the list of layers, so as written it's a
no-op and the Info column is never cleared. Change this to what
is intended, and clear the column if this is the first HTTP/3 STREAM
in the frame by checking proto_get_layer_num for proto_http3 instead.