pcapng: remove a no-longer-needed union.
A custom_binary_opt_t will never contain any structures specific to a
particular PEN - that gets in the way of arbitrary custom options.
Remove the union.

While we're at it, remove PEN_NFLX from wiretap/wtap_opttypes.h; either
we should have a generic header - in wsutil, as it'd be used both by
libwiretap and libwireshark - containing PENs, or we should put each PEN
in the libwiretap header for that particular PEN's custom data.  Do the
latter, for now.