Register sysdig pcapng blocks via register_pcapng_block_type_handler
Move to the idea of having less "internal" block types handled and create the idea of "built in" block types that are registered the same way as plugins, they are just done "first", so they get priority.

Treat the sysdig pcapng blocks as if they were a plugin.  This helps shape the API of registering pcapng block handlers