tap-iostat: Fix a possible floating point exception crash
The -z io,stat tap does not necessarily have a comma after
the interval value (in the default case of no filter, only
counting frames and bytes).

g_strstr_len, like strstr and strchr, does not return a pointer
to the null terminator when it fails to find a value. Since it
returns a NULL pointer (numeric value 0), this can cause invl_len
to be assigned to a large negative number in the case of

tshark -z io,stat,1.00

or similar. Since strchrnul is a GNU extension, implement the logic
using strpbrk and strcspn. Also make the intermediate variable
unsigned, since it's assigned to an unsigned variable.

(backported from commit 2544a1251b09c9b9790ab3febc3849b9d084b879)