tap-iostat: Get value in a separate static function
Get the value for each item in a separate function, and add the
borders for the text table around it in the main function. This
will make it easier to implement other output modes like CSV.
Note that the single entries that have both frame and byte counts
will complicate that.