GitLab CI: Update macOS installer to Qt 6.8.3
Since the official macOS package libraries are targeting macOS 12
as a development target, let's move to Qt 6.8 (like the Windows
package) which has a minimum deployment target of macOS 12 as well.
Qt 6.5.3 is long unsupported at this point.

This is necessary for #19328 and also should resolve #20323 according
to the Qt bug tracker.