PROFINET: Fix protectionMode bitmask upon retrieval
The bitmask in the hfi definition is 0x01, there are only two
values (0 and 1) in the value string, and the code paths assume
that the value can only be 0 or 1 (and fail to initialize a variable
if it's otherwise.) So mask by 0x1.

Also fix some use of tabs in newly added hfi definitions, different
from the alignment of all other hfis.

Fix #20540