PROFINET: Fix more leaks
Add a wmem_allocator scope member to the ParserState object, and
use that to allocate its lists and strings, instead of strdup for
the strings. This eliminates the leaks.

This still isn't optimal memory use. The same GSD file is parsed
for each conversation, and the parsing state uses file scope.
Either it should be parsed in pinfo->pool scope and the information
copied to the file scoped stationInfo (copying the strings, as they
would be in different scopes), or (ideally, as opening a file and
parsing it can be slow) it should be parsed once and copied to the
stationInfo for each conversation as needed (in that case the strings
might be able to be shared if they are the same scope.)

Fixup 6d1dab030a25e021eb2b01f934c801c4665c7521