tap-macltestat: add finish routine
Add a finish function to the call of `register_tap_listener()` which
frees the memory allocated by the tap. This includes the tap data
allocated upon registration, as well as the list of UE rows.

Add code to the reset function which frees the memory allocated for UE
rows, instead of simply leaking them by resetting the list pointer to
`NULL`.

Ping #20432