tap-oran: address memory management issues
Add a finish function to the call of `register_tap_listener()` which
frees the memory allocated by the tap. This includes the tap data
allocated upon registration, as well as the list of flows.

Add code to the reset function which frees the memory allocated for
flows, instead of simply leaking them by resetting the list pointer to
`NULL`.

In `oran_stat_packet()`, sort and iterate over `hs->flow_list` directly,
bypassing the temporary variable `list`. All operations on `GList`s,
including `g_list_sort()`, return a pointer to the new start of the
list; thus, without this change, only `list` is valid and
`hs->flow_list` becomes a stale pointer after the sort.

Ping #20432