tap-protocolinfo: minor memory management tweaks
Add a finish function to the call of `register_tap_listener()` which
frees the memory allocated by the tap. This is just the tap data which
is allocated when the listener is registered.

When registering the listener, free our copy of the filter string.
`register_tap_listener()` takes its own copy of it, so we're leaking our
copy if we don't free it.

When registering the listener, replace two-stage
`g_malloc()`+`g_strlcpy()` with a single call to `g_strndup()`. Note the
minor semantic difference between the two: [`g_strlcpy()`][CPY] copies
"At most `dest_size`-1 characters" whereas [`g_strndup()`][DUP] copies
"the first `n` bytes of a string". (Both guarantee a null terminator on
the new copy.)

[CPY]: https://docs.gtk.org/glib/func.strlcpy.html
[DUP]: https://docs.gtk.org/glib/func.strndup.html

Ping #20432