tap-rtspstat: improve memory management
Add a finish function to the call of `register_tap_listener()` which
frees the memory allocated by the tap. This includes the tap data
allocated upon registration, as well as the `GHashTable`s holding
request and response statistics.

Replace `g_hash_table_new()` calls with `g_hash_table_new_full()` to
provide callback functions for freeing the keys and values of hash
entries as they are being removed.

* The keys for `hash_responses` are numbers stored using
  `GINT_TO_POINTER()` so no callback is needed. The values are created
  with `g_new()` so `g_free` is a sufficient callback.
* The keys for `hash_requests` are strings, which are also pointed to by
  the a member of the `g_new()`-allocated values. `g_free` as the
  callback for both the key and value takes care of it, but this may be
  non-obvious for future maintainers.

Ping #20432