tap-sctpchunkstat: add finish routine
Add a finish function to the call of `register_tap_listener()` which
frees the memory allocated by the tap. This includes the tap data
allocated upon registration, as well as the list of SCTP statistics
elements.

Register the existing function `sctpstat_reset()` as the reset handler
for the tap, so it gets called as wireshark/tshark need it, rather than
just once when the listener is created.

Ping #20432