Have wiretap provide the name for the type of a recored.
This removes two places that need to know about specific custom block
PENs and that would need to change if we add another record type.

Also, don't use the block type name in the top-level "frame" protocol
item.  The number following it is an ordinal number of the record within
the file, independent of the record type - it's not as if "Frame 73" is
necessarily the 73rd packet in the file, as there might have been custom
blocks or system journal export items or system call reports or..., and
the same applies for any *other* record type.

Use the "frame" protocol for the top-level item, even for Netflix custom
blocks; 1) all the items *under* it are for that protoocol and 2) the
dissection of the custom block data use the "bblog" protocol, so that
protocol is present in the dissection.

The name really *should* be "Record", but, for now, we use "Frame" in
case anybody looks for "Frame" at the beginning of a line.

The *protocol" should also be "record", but we'd need to have "frame" as
an alias, and "frame.*" as an alias for all "record.*" fields, for
backwards compatibility.