pcapng-netflix-custm: don't use sizeof for the tcpinfo option size.
struct nflx_tcpinfo has 64-bit integral type values in it, but the sum
of the sizes of the elements plus internal padding is *not* a multiple
of 8, so, on a platform on which 64-bit integral type values are aligned
on an 8-byte boundary - i.e., on all 64-bit platforms on which we run,
probably meaning on the majority of machines on which Wireshark is run
these days, especially given that we don't support 32-bit Windows or
macOS any more - it will have 4 bytes of unnamed padding at the end.

The custom option data in capture files does *not* necessarily include
the unnamed padding.

Instead, explicitly define OPT_NFLX_TCPINFO_SIZE as the size consumed by
the *data* in the option - not padding added on 64-bit platforms - and
compare against *that* in a ws_assert().

This isn't an issue if asserts aren't compiled in, but will probably
blow up on 64-bit platforms if options are compiled in.