Qt: Update DisplayFilterCombo on theme change for all OSes
QEvent::ApplicationPaletteChange is sent to top-level windows, which
propagate QEvent:PaletteChange to children, but not to children that
use style sheets. We have child widgets that use style sheets but
vary appearance based on the application palette (generally whether
it's considered Light or Dark.)

For DisplayFilterCombo, the shading of the downward arrow varies
on light and dark mode for every OS, not just macOS, so it needs to
be sent in all cases.