dfilter: For contains, implicitly convert field with value strings
Like "matches", implicitly convert fields with value strings
to their value strings. (Byte arrays never have value strings, and
contains only compares strings or byte arrays, so this does not
cause ambiguity.) For now, only apply to the LHS, but this can
probably be extended to a field on the RHS.

(Separately, "contains", as it doesn't commute, might be an operation
where it could make sense to allow literals on the left and the field
on the right.)

Cf. 6275d509e1c2d7b606adf22c4976c8d4753046f8