wslua: Fix some leaks on errors with gcrypt
Avoid leaking when there are errors (which is tested in the gcrypt.lua test
script) so that the tests pass when built with ASAN.

Also fix the GcryptCipher.open related tests. There was an
unnecessarily defined function expecting to be passed in a
GcryptCipher object, but none of the tests passed in such a
parameter (since the test opens it), so the wrong parameters
were passed to the open function. (The ones that failed still
failed, just differently.)

Fixup 0d74b79e2886252eb9a80ad749de83c161212f68