dftest: Convert filter to UTF-8 on Windows
Have dftest include cli_main so that the filter (a command line
argument) is converted from UTF-16 to UTF-8, because that's what
happens when using filters in Wireshark, TShark, or the other
programs. It also prevents some oddities when trying to output
to the screen the filter.

Ping #20501

Before:

.\run\RelWithDebInfo\dftest.exe 'tcp.payload contains \"é\"'
Filter:
 tcp.payload contains "Instructions:
 0000 READ_TREE        tcp.payload      -> R0
 0001 IF_FALSE_GOTO    3
 0002 ANY_CONTAINS     R0 contains e9:
 0003 RETURN

After:

.\run\RelWithDebInfo\dftest.exe 'tcp.payload contains \"é\"'
Filter:
 tcp.payload contains "é"

Instructions:
 0000 READ_TREE        tcp.payload      -> R0
 0001 IF_FALSE_GOTO    3
 0002 ANY_CONTAINS     R0 contains c3:a9
 0003 RETURN