frame: change the way we avoid unnecessary dissection steps.
Test whether "frame" or any "frame.*" fields are referenced and save it
in a Boolean variable.  THen just have one code path and test that
variable as appropriate.

This means we only have one place where we add expert infos, and we make
sure it's done *outside* the tests of that variable.

This should make the dissection code a bit easier to maintain.