frame: do all the Netflix custom option stuff in one place.
This makes it easier to move it to a routine called through a "custom
option for a particular PEN" dissector table.

Move other stuff in that "modify pinfo" section down as well to the next
switch-on-record-type statement, add a case for an un known record type,
and get rid of the other switch statement.

Don't use DISSECTOR_ASSERT() in that switch statement, as it throws an
exception, and the TRY that catches the exception is later. Just report
an error in the summary and detail panes.