Reduce the specialized handling of Netflix custom options.
Don't have specialized routines in libwiretap for fetching and setting
Netflix custom options; just use the standard APIs for handling custom
options.

In the frame dissector, instead of a wired-in check for Netflix custom
options, just loop over all options and process the custom ones - for
now, have a routine to handle the Netflix ones and ignore the others.

This is a first step towards generalizing the handling of custom options
in libwireshark.