frame: have a dissector table for custom binary options.
Move the Black Box Log option dissection code to
epan/dissectors/packet-bblog.c, and register it in the aforementioned
dissector table.

This removes the last of the code in packet-frame.c that knows anything
abut Netflix custom options, and allows it to handle all custom binary
options, as long as they register a handler in the aforementioned
dissector table.

Make the code in packet-frame.c that generates the top-level frame
protocol tree item start ut with common code to generate the entry,
followed by initial per-record-type code tha appends the relevant data
to that enry.