GitLab CI: Have the Commit Check pre-commit run on the first commit
The pre-commit tool, as a hook, expects to run *before* a commit
happens, so it compares HEAD to the working tree/index. When run
after a commit, it expects to be passed the ancestor of the new
commit, i.e. HEAD~ not HEAD.

It also always checked multiple commits, but produced a single output
result for the collective total of commits instead of splitting them.
As it is now, passing in several commits ends up producing reports
that are subsets of each other (by comparing each commit to the current
tree/HEAD, instead of comparing each commit to its immediate parent.)

So currently the Commit Check is not running the pre-commit on the
first commit of a MR (which means that it's not running it at all
in the typical case of a 1 commit MR.)