tap-sipstat: improve memory management
Add a finish function to the call of `register_tap_listener()` which
frees the memory allocated by the tap. This includes the tap data
allocated upon registration, as well as the filter string and the
`hash_requests` and `hash_responses` hash tables.

Create those tables using `g_hash_table_new_full()`, passing `g_free`
as the destroy functions for the keys and values, so GLib will free the
memory we allocate for them.

Ping #20432