Improve handling of unwritable records.
Provide an info string with WTAP_ERR_UNWRITABLE_REC_TYPE errors, and
reprt the string. Have the string indicate what the particular problem
is.

For file-type-specific events and reports, include the file type with
in the record header, so that the writer can reject file-type-specifc
events and reports for a different file type.  Do so for pcapng.

Have the routine for setting up records for file-type-specific events
and reports take the file type and record type as arguments, and
initialize those fields.