Qt: CaptureInfoDialog, if enabled, should show when capturing on startup
If the "-k" option is specified and the Capture Info Dialog is enabled,
it should show on startup. To do that, the main window needs to set
itself appropriately in the void * in info_data (as done manually for
starting a capture.) Do that on instantiation of the main window.