Qt: Avoid using QTextCodec in simple_dialog
QTextCodec is part of the Qt5 Core Compat module, so reduce usage
of it.

QString::fromLocal8Bit() is available in both Qt5 and Qt6 and does
the same thing.

On Qt 5, QString::fromLocal8Bit() uses QTextCodec::codecForLocale(),
whereas on Qt 6 it is the same as QString::fromUtf8(), but Qt 6
always forces the process locale to use UTF-8 on UN*X anyway so it
gives the same result:
https://github.com/qt/qtbase/commit/419f42903154489e2dace8d1eba183036275448b

The other cases of QTextCodec are harder to eliminate; Qt drastically
reduced the number of included codecs with the Qt6 alternative unless
building with ICU. The Qt Company's binary packages for Windows aren't
built with libicu, so that approach to preserving GUI support for
Shift-JIS, KOI8-R, Windows-1252 etc. would make development more
challenging.