SSH: Handle defragmentation and truncated packets
Support defragmenation of encrypted SSH packets under the ciphers
we handle. (The file in #20332 has fragmentation).

Support handling of truncated packets with ChaCha20-Poly1035 and
AES-GCM we will be able to dissect subsequent packets after truncated
frames. It ought to be possible to do so for AES-CTR mode as well;
add a relevant comment.

Add some other comments.

Fix #20332 (at least up to a certain size)

Ping #16054