pcapng: clean up padding-to-4-bytes code.
Add a new wsutil header with macros that compute the amount of padding
needed to round up to a 4-byte boundary (without any conditional tests)
- it ends up being a negate and an AND.

Use that in pcapng code. (Next step: use it everywhere else.)

Have an inline function that takes a number of bytes of padding to
write, between 0 and 3, and writes out that many bytes worth of zeroes.
Use it.