SSH: Check payload lengths and take a subset tvb
Check the payload length of decrypted packets (calculated from
the packet length and the padding length), and take a subset
tvb. This is necessary because some payload types do not have
a fixed number of items, but can have variable length arrays
(e.g., the hostkeys-00@openssh.com SSH_MSG_GLOBAL_REQUEST), so
the total payload length is needed to know when to stop.

Do some more checking in the case of impossible results that
would cause overflow. That shouldn't happen in a decrypted
packet where the MAC passed; if it does, something went badly
wrong.

Add a few comments and cleanup the style to be more consistent.