smb: fix handling of lists of extended attributes.
Fix the routine that handles SMB_FEA_LIST:

* Take into account that the size of the SMB_FEA_LIST, in bytes,
includes not only all the SMB_FEA items in the list, but the list size
itself, and to ensure that the list size is large enough to include the
list size and report an error if it isn't.

* Add checks to make sure we don't go beyond the list size.

* Rename it to dissect_smb_fea_list(), to make it clearer what it does,
rather than giving an item number in some document that isn't even the
item that deescribes an SMB_FEA_LIST or an SMB_FEA.

Add a routine to handle SMB_GEA_LIST.

Use those routines where appropriate.