pcapng: clean up some padding.
For SPBs, there's no need to recompute the padding after the packet data
right before we skip over it; it's already been computed and saved into
a variable earlier in the code.

For options, use WS_ROUNDUP_4() to round up the size of an option to a
multiple of 4 bytes.