pcapng: clean up the NRB-reading code.
Just have to_read count the bytes remaining to be read, which makes the
tests for "do we have enough bytes left? a bit simpler.

Calculate the NRB padding up front and save it in a variable for later
use.

Move the "skip the padding" outside the switch statement - it's the same
for all record types.