pcapng: pass the block *content* length to block handlers.
Do the check for "is this block large enough for a block header and
footer" in common code, before a handler is called.

Pass the block *content* length to block handlers, so they don't need to
know about the header+footer length.

Move MIN_BLOCK_SIZE to wiretap/pcapng.c, as it's no longer needed
outside that file.

Get rid of minimum block size definitions that don't need to be used.

Make the flow of the PB/EPB processing code and of the SPB processing
code more like one another.

Add some additional length checks for DSBs.

Remove redundant "look for a registered block handler" code that was
only called if we didn't find one.