tap-wspstat: fix hash and add finish callback
When registering the tap, also register a "finish" callback to clean up
memory used by the tap. Populate this function with the code that was
being used to free the allocated tap data in the event of a failed
registration.

Remove `wsp_free_hash_table()`; replace `g_hash_table_new()` with
`g_hash_table_new_full()` instead, to let GLib handle cleaning it up.

The hash was being initialized with GLib-allocated keys, but hash misses
were being added with `GINT_TO_POINTER()` values as keys. Correct the
initialization to also use `GINT_TO_POINTER()`. (This actually results in
an improvement in the output: when run on [`wap_google.pcap`][1], the
returned status codes are now reported as "200 OK" rather than
"(null)".)

Ping #20432

[1]: https://gitlab.com/wireshark/wireshark/-/wikis/SampleCaptures#wap-protocol-family