Qt: Preserve simple dialog secondary message when queueing messages
Creating the SimpleDialog and then setting the secondary text later
doesn't work when the app is still initializing and queuing all the
messages to show later. It ends up dropping the secondary message.

Pass the secondary message as an optional parameter to the constructor.

Do the trimming and cleanup of the primary and secondary messages before
queueing them. Only create the MessagePair struct if we're queueing
them, and construct it in place.