packet-rtsp: memory management tweaks
Remove `rtsp_stat_info` as a static variable. Instead, define it in
`dissect_rtspmessage()` and pass it as a parameter to the one other
function that needs it, `is_rtsp_request_or_reply()`.

Also add a `wmem_allocator_t *pool` parameter to this function, for
specifying the scope in which to allocate a string for `rtsp_stat_info`,
instead of using `wmem_packet_scope()`.

These changes are small steps toward potential future concurrency
support in Wireshark as envisioned
[here](https://www.wireshark.org/lists/wireshark-dev/202107/msg00052.html).