ssh: Add heuristic dissector
Add a SSH heuristic dissector that looks for the Version Identification
String.

If we have heuristics, we can't just use match_uint (which isn't set
when heuristics or conversation matching are used for the dissector.)

There should be some additional changes later to handle the rare case
where the server version string comes first, or if the client version
string is missing. The initial strings and SSH_MSG_KEXINIT are equivalent
in either direction; it is not until the key exchange actually starts that
server versus client can be determined.