ssh: Provide static functions for determining in_response
Provide functions for determining if the peer_data or the
current packet is from the server. The technique of looking at
pinfo->match_uint won't work for heuristic dissectors, if implemented,
and don't work for SCTP if the PPI was used to match and the port
is non-standard, so this separates it out and only does that test
in one place.

Also factor out some common code and eliminate some unnecessary moving
through multiple pointers to get back to the same peer_data we already
had.