BTATT: Use the remote BD ADDR when tracking MTUs
The ATT_MTU is negotiated between a particular server and client (and is
the same in both directions - "A device that is acting as a server and
client at the same time shall use the same value for Client Rx MTU and
Server Rx MTU," Bluetooth Core Specification Volume 3 Part F 3.2.8
Exchanging MTU size.)

Therefore, the MTU size is only relevant for a particular pair of
addresses. Use the remote BD ADDR when storing and looking up the
MTU.

It might make sense in the future to create a conversation instead of
having many trees that all use similar keys.

Ping #20537